Started work on the alt analysis system and finalized post analysis

+++ PycharmProjects/NRPSQaunification/TwoSeqCompare.py
@@ -8,6 +8,8 @@ from io import StringIO
 import BLASTWriter
 import os
 search_database = "nucleotide"
+ma_dir = "Matrix"
+alt_dir = "Alternative Analysis"
 root_dir = "NRPSRoot"
 dat_dir = "NRPSData"
 nrps_dir= "NRPS"
@@ -31,21 +33,39 @@ def compare(subjct, records):
     sbjct = SeqRecord(SeqIO.read(os.path.join(ana_dir, os.path.join(nrps_dir, os.path.join(dat_dir, os.path.join(
                 gen_dir, subjct.get() + ".gbk")))), format="gb"), id="sub")
     for record in records:
-        recrd = SeqIO.read(os.path.join(root_dir, record[:len(record) - 4]) + ".gbk", format="gb")
-        rcrd = SeqRecord(recrd, id="rec")
-        SeqIO.write(rcrd.seq, "rec.fasta", "fasta")
-        SeqIO.write(sbjct.seq, "sub.fasta", "fasta")
-        result_handle = NcbiblastnCommandline(query="rec.fasta", subject="sub.fasta", outfmt=5)()[0]
-        os.remove("sub.fasta")
-        os.remove("rec.fasta")
-        blast_result_record = NCBIXML.read(StringIO(result_handle))
-        # Print some information on the result
-        i = 0
-        for alignment in blast_result_record.alignments:
-            for hsp in alignment.hsps:
-                BLASTWriter.create_blast_files(open(os.path.join(ana_dir, os.path.join(nrps_dir, os.path.join(dat_dir,
-                                                   os.path.join(dat_stan_dir, subjct.get())))), 'w'),
-                                                       BLASTWriter.write_blast_standard(i, alignment, hsp))
+        if record is not None:
+            recrd = SeqIO.read(os.path.join(root_dir, record + ".gbk"), format="gb")
+            rcrd = SeqRecord(recrd, id="rec")
+            SeqIO.write(rcrd.seq, "rec.fasta", "fasta")
+            SeqIO.write(sbjct.seq, "sub.fasta", "fasta")
+            result_handle = NcbiblastnCommandline(query="rec.fasta", subject="sub.fasta", outfmt=5)()[0]
+            os.remove("sub.fasta")
+            os.remove("rec.fasta")
+            blast_result_record = NCBIXML.read(StringIO(result_handle))
+            # Print some information on the result
+            i = 0
+            csvGenerator.create_dir(os.path.join(dat_dir, os.path.join(dat_stan_dir, subjct.get())), nrps_dir)
+            k = 0
+
+            for alignment in blast_result_record.alignments:
+                csvGenerator.create_dir(os.path.join(dat_dir, os.path.join(dat_stan_dir, os.path.join(subjct.get(),
+                                            alignment.title + " - " + str(k)))),
+                                                nrps_dir)
+                j = 0
+                for hsp in alignment.hsps:
+                    BLASTWriter.create_blast_files(open(os.path.join(ana_dir, os.path.join(nrps_dir, os.path.join(
+                                                        dat_dir, os.path.join(dat_stan_dir, os.path.join(subjct.get(),
+                                                            os.path.join(alignment.title + " - " + str(k),
+                                                                alignment.title + " - " + str(k) +
+                                                                    " - hsp" + str(j))))))), 'w'),
+                                                                        BLASTWriter.write_blast_standard(i, alignment,
+                                                                            hsp))
+                j += 1
                 i += 1
-                return open(os.path.join(ana_dir, os.path.join(nrps_dir, os.path.join(dat_dir,
-                                                   os.path.join(dat_stan_dir, subjct.get())))), 'r')
+
+            k += 1
+            return os.path.join(ana_dir, os.path.join(nrps_dir, os.path.join(dat_dir,
+                                                                              os.path.join(dat_stan_dir,
+                                                                                           subjct.get()))))
+def alt_compare(subjct, record):
+    csvGenerator.create_dir()
